+++ chainstate/test-suite/src/tests/double_spend_tests.rs
@@ -48,7 +48,7 @@ use crypto::random::SliceRandom;
 #[trace]
 #[case(Seed::from_entropy())]
 fn spend_output_in_the_same_block(#[case] seed: Seed) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut tf = TestFramework::default();
 
         let mut rng = make_seedable_rng(seed);
@@ -81,7 +81,7 @@ fn spend_output_in_the_same_block(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 fn spend_output_in_the_same_block_invalid_order(#[case] seed: Seed) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut tf = TestFramework::default();
 
         let mut rng = make_seedable_rng(seed);
@@ -122,7 +122,7 @@ fn spend_output_in_the_same_block_invalid_order(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 fn double_spend_tx_in_the_same_block(#[case] seed: Seed) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut tf = TestFramework::default();
 
         let mut rng = make_seedable_rng(seed);
@@ -169,7 +169,7 @@ fn double_spend_tx_in_the_same_block(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 fn double_spend_tx_in_another_block(#[case] seed: Seed) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut tf = TestFramework::default();
 
         let mut rng = make_seedable_rng(seed);
@@ -209,7 +209,7 @@ fn double_spend_tx_in_another_block(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 fn overspend_single_output(#[case] seed: Seed) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut tf = TestFramework::default();
 
         let mut rng = make_seedable_rng(seed);
@@ -240,7 +240,7 @@ fn overspend_single_output(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 fn overspend_multiple_outputs(#[case] seed: Seed) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut tf = TestFramework::default();
         let mut rng = make_seedable_rng(seed);
 
@@ -293,7 +293,7 @@ fn overspend_multiple_outputs(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 fn duplicate_input_in_the_same_tx(#[case] seed: Seed) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut tf = TestFramework::default();
 
         let mut rng = make_seedable_rng(seed);
@@ -346,7 +346,7 @@ fn duplicate_input_in_the_same_tx(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 fn same_input_diff_sig_in_the_same_tx(#[case] seed: Seed) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut tf = TestFramework::default();
 
         let mut rng = make_seedable_rng(seed);
@@ -404,7 +404,7 @@ fn same_input_diff_sig_in_the_same_tx(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 fn duplicate_tx_in_the_same_block(#[case] seed: Seed) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut tf = TestFramework::default();
 
         let mut rng = make_seedable_rng(seed);
@@ -434,7 +434,7 @@ fn duplicate_tx_in_the_same_block(#[case] seed: Seed) {
 #[trace]
 #[case(Seed::from_entropy())]
 fn duplicate_odd_tx_in_the_same_block(#[case] seed: Seed) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut tf = TestFramework::default();
 
         let mut rng = make_seedable_rng(seed);
