+++ chainstate/test-suite/src/tests/helpers/in_memory_storage_wrapper.rs
