+++ chainstate/test-suite/src/tests/tx_verification_simulation.rs
@@ -20,7 +20,7 @@ use chainstate_test_framework::TxVerificationStrategy;
 #[trace]
 #[case(Seed::from_entropy(), 20, 50)]
 fn simulation(#[case] seed: Seed, #[case] max_blocks: usize, #[case] max_tx_per_block: usize) {
-    ::utils::concurrency::model(move || {
+    utils::concurrency::model(move || {
         let mut rng = make_seedable_rng(seed);
         let mut tf = TestFramework::builder()
             .with_tx_verification_strategy(TxVerificationStrategy::Randomized(seed))
